Make profiling sampling a stub package
