+++ stdlib/profiling/sampling/__init__.pyi
@@ -1,0 +1,17 @@
+from .collector import Collector as Collector
+from .gecko_collector import GeckoCollector as GeckoCollector
+from .heatmap_collector import HeatmapCollector as HeatmapCollector
+from .jsonl_collector import JsonlCollector as JsonlCollector
+from .pstats_collector import PstatsCollector as PstatsCollector
+from .stack_collector import CollapsedStackCollector as CollapsedStackCollector
+from .string_table import StringTable as StringTable
+
+__all__ = (
+    "Collector",
+    "PstatsCollector",
+    "CollapsedStackCollector",
+    "HeatmapCollector",
+    "GeckoCollector",
+    "JsonlCollector",
+    "StringTable",
+)
