+++ stdlib/profiling/sampling/gecko_collector.pyi
@@ -1,0 +1,8 @@
+from _typeshed import StrOrBytesPath
+
+from .collector import Collector, _StackFrames, _Timestamps
+
+class GeckoCollector(Collector):
+    def __init__(self, sample_interval_usec: int, *, skip_idle: bool = False, opcodes: bool = False) -> None: ...
+    def collect(self, stack_frames: _StackFrames, timestamps_us: _Timestamps = None) -> None: ...
+    def export(self, filename: StrOrBytesPath) -> None: ...
