+++ stdlib/profiling/sampling/heatmap_collector.pyi
@@ -1,0 +1,20 @@
+from _typeshed import StrOrBytesPath
+from collections.abc import Sequence
+
+from .collector import Collector, _Frame, _StackFrames, _Timestamps
+
+class HeatmapCollector(Collector):
+    FILE_INDEX_FORMAT: str
+    def __init__(self, sample_interval_usec: int, *, skip_idle: bool = False) -> None: ...
+    def collect(self, stack_frames: _StackFrames, timestamps_us: _Timestamps = None) -> None: ...
+    def export(self, output_path: StrOrBytesPath) -> None: ...
+    def process_frames(self, frames: Sequence[_Frame], thread_id: int, weight: int = 1) -> None: ...
+    def set_stats(
+        self,
+        sample_interval_usec: int,
+        duration_sec: float,
+        sample_rate: float,
+        error_rate: float | None = None,
+        missed_samples: float | None = None,
+        **kwargs: object,
+    ) -> None: ...
