+++ stdlib/profiling/sampling/stack_collector.pyi
@@ -1,0 +1,33 @@
+from _typeshed import StrOrBytesPath
+from abc import ABCMeta
+from collections.abc import Sequence
+
+from .collector import Collector, _Frame, _StackFrames, _Timestamps
+
+class StackTraceCollector(Collector, metaclass=ABCMeta):
+    def __init__(self, sample_interval_usec: int, *, skip_idle: bool = False) -> None: ...
+    def collect(self, stack_frames: _StackFrames, timestamps_us: _Timestamps = None) -> None: ...
+    def process_frames(self, frames: Sequence[_Frame], thread_id: int, weight: int = 1) -> None: ...
+
+class CollapsedStackCollector(StackTraceCollector):
+    def __init__(self, sample_interval_usec: int, *, skip_idle: bool = False) -> None: ...
+    def process_frames(self, frames: Sequence[_Frame], thread_id: int, weight: int = 1) -> None: ...
+    def export(self, filename: StrOrBytesPath) -> None: ...
+
+class FlamegraphCollector(StackTraceCollector):
+    def __init__(self, sample_interval_usec: int, *, skip_idle: bool = False) -> None: ...
+    def collect(self, stack_frames: _StackFrames, timestamps_us: _Timestamps = None) -> None: ...
+    def set_stats(
+        self,
+        sample_interval_usec: int,
+        duration_sec: float,
+        sample_rate: float,
+        error_rate: float | None = None,
+        missed_samples: float | None = None,
+        mode: int | None = None,
+    ) -> None: ...
+    def export(self, filename: StrOrBytesPath) -> None: ...
+    def process_frames(self, frames: Sequence[_Frame], thread_id: int, weight: int = 1) -> None: ...
+
+class DiffFlamegraphCollector(FlamegraphCollector):
+    def __init__(self, sample_interval_usec: int, *, baseline_binary_path: StrOrBytesPath, skip_idle: bool = False) -> None: ...
