+++ stdlib/profiling/sampling/string_table.pyi
@@ -1,0 +1,5 @@
+class StringTable:
+    def intern(self, string: object) -> int: ...
+    def get_string(self, index: int) -> str: ...
+    def get_strings(self) -> list[str]: ...
+    def __len__(self) -> int: ...
